Fix warning in Android position plugin
qtlocation/src/plugins/position/android/src/qgeopositioninfosource_android.cpp:226:
    warning: implicit conversion loses integer precision: 'qint64' (aka 'long long') to 'const int'

+++ src/plugins/position/android/src/qgeopositioninfosource_android.cpp
@@ -223,7 +223,7 @@ void QGeoPositionInfoSourceAndroid::requestTimeout()
         const QGeoPositionInfo info = queuedSingleUpdates[i];
 
         //anything newer by 20s is always better
-        const int timeDelta = best.timestamp().secsTo(info.timestamp());
+        const qint64 timeDelta = best.timestamp().secsTo(info.timestamp());
         if (abs(timeDelta) > 20) {
             if (timeDelta > 0)
                 best = info;
